+++ examples/json/json.go
@@ -56,7 +56,7 @@ func main() {
 		terminal.Rune('}', "}"),
 	)
 
-	value = combinator.Memoize("VALUE", h, combinator.Any("value",
+	value = h.Memoize(combinator.Any("value",
 		terminal.String(),
 		terminal.Integer(),
 		terminal.Float(),